Rename file. Support matching against partial alphabet matches
